Feat/profile (#17)
* chore: update workflows

* refactor: refactor graphql folder

* feat: add image constraint
    - only png and jpeg are allow in upload
    - search for image in route /image to judge it exist or not

* feat: add profile, user, author and authorbookrel schemas

* fix: add date format for date_of_birth in profile

* fix: last_name is not unique

* chore: refactor code for graphql, models, services

* chore: update image get route

+++ src/models/schemas/Book.Schema.ts
@@ -1,9 +1,9 @@
-import { Schema } from 'mongoose';
+import { Schema, Types } from 'mongoose';
 import moment from 'moment';
-import { isDate } from 'util/types';
-import { GenreBookRelModel, ImageGridFsBucket } from '..';
-import { BookData } from '../../types/models';
+import { BookData } from 'src/types';
 import { createMongoObjectIdFromString } from 'src/utils';
+import { isDate } from 'moment';
+import { AuthorBookRelModel, GenreBookRelModel, ImageGridFsBucket } from '..';
 
 const BookSchema = new Schema<BookData>(
   {
@@ -14,7 +14,7 @@ const BookSchema = new Schema<BookData>(
     },
     description: String,
     rating: Number,
-    cover_image: String,
+    cover_image: Types.ObjectId,
     publish_date: {
       type: Date,
       validate: {
@@ -37,13 +37,20 @@ BookSchema.pre<BookData>('deleteOne', async function (next) {
 
   // clear book in genres when book is deleted
   GenreBookRelModel.deleteMany({
-    bookId: bookCurrent._id,
+    bookId: createMongoObjectIdFromString(bookCurrent._id.toString()),
+  }).session(await GenreBookRelModel.startSession());
+
+  AuthorBookRelModel.deleteMany({
+    bookId: createMongoObjectIdFromString(bookCurrent._id.toString()),
   }).session(await GenreBookRelModel.startSession());
 
   // clear Image in GridFs when delete
-  ImageGridFsBucket.delete(
-    createMongoObjectIdFromString(bookCurrent.cover_image.toString()),
-  );
+  if (bookCurrent.cover_image) {
+    ImageGridFsBucket.delete(
+      createMongoObjectIdFromString(bookCurrent.cover_image.toString()),
+    );
+  }
+
   next();
 });
